% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{adapt_df_to_opt_windows}
\alias{adapt_df_to_opt_windows}
\title{Filter a data frame according to the optimization window start hour}
\usage{
adapt_df_to_opt_windows(df, window_start_hour)
}
\arguments{
\item{df}{data frame, first column named \code{datetime} of type \code{datetime}}

\item{window_start_hour}{integer, hour to start the optimization window}
}
\value{
tibble
}
\description{
Filter a data frame according to the optimization window start hour
}
