% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{adapt_dttm_seq_to_opt_windows}
\alias{adapt_dttm_seq_to_opt_windows}
\title{Cut the dateteime sequence according to the optimizion window start hour and length}
\usage{
adapt_dttm_seq_to_opt_windows(
  dttm_seq_original,
  window_start_hour,
  window_length
)
}
\arguments{
\item{dttm_seq_original}{datetime sequence of the original timeseries data}

\item{window_start_hour}{integer, hour to start the optimization window}

\item{window_length}{integer, number of data points of the optimization window (not in hours)}
}
\value{
datetime sequence
}
\description{
Cut the dateteime sequence according to the optimizion window start hour and length
}
