% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{add_battery}
\alias{add_battery}
\title{Battery optimal charging/discharging profile}
\usage{
add_battery(
  w,
  G,
  L,
  Bcap,
  Bc,
  Bd,
  SOCmin = 0,
  SOCmax = 100,
  SOCini = NULL,
  SOCend = NULL,
  up_to_G = TRUE,
  window_length = NULL
)
}
\arguments{
\item{w}{numeric between \code{0} and \code{1}, being the weight of the flexibility potential}

\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the window beginning}

\item{SOCend}{numeric, required State-of-Charge at the window end}

\item{up_to_G}{logical, whether to limit the flexible EV demand up to renewable Generation}

\item{window_length}{integer, window length. If \code{NULL}, the window length will be the length of \code{G}}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile
}
