% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{add_battery_window}
\alias{add_battery_window}
\title{Battery optimal charging/discharging profile (just a window)}
\usage{
add_battery_window(w, G, L, Bcap, Bc, Bd, SOCmin, SOCmax, SOCini)
}
\arguments{
\item{w}{numeric between \code{0} and \code{1}, being the weight of the flexibility potential}

\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile (just a window)
}
