% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_extra_days}
\alias{add_extra_days}
\title{Add an extra day at the beginning and the end of datetime sequence
using the last and first day of the data}
\usage{
add_extra_days(df)
}
\arguments{
\item{df}{data frame, first column named \code{datetime} of type \code{datetime}}
}
\value{
tibble
}
\description{
Add an extra day at the beginning and the end of datetime sequence
using the last and first day of the data
}
