% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{add_storage_level}
\alias{add_storage_level}
\title{Add \code{storage} column to time-series tibble}
\usage{
add_storage_level(df, power, init = 0, loss = 0)
}
\arguments{
\item{df}{tibble, being the first column \code{datetime} variable}

\item{power}{numeric vector, being positive when charging and negative when discharging}

\item{init}{numeric, initial storage level (in energy units, not \%)}

\item{loss}{numeric, the hourly storage loss in percentage (\%/hour), passed to \code{get_storage_losses} function}
}
\value{
tibble, with extra column \code{storage}
}
\description{
Add \code{storage} column to time-series tibble
}
