% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{aggregate_timeseries}
\alias{aggregate_timeseries}
\title{Aggregate multiple timeseries columns to a single one}
\usage{
aggregate_timeseries(df, varname, omit = NULL)
}
\arguments{
\item{df}{data.frame or tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{varname}{character, name of the aggregation column}

\item{omit}{character, name of columns to not aggregate}
}
\value{
tibble
}
\description{
The first column \code{datetime} will be kept.
}
\examples{
# Example data set with 2 identical building profiles
df <- dplyr::select(
  energy_profiles, datetime, building1 = building, building2 = building
)
head(df)

# Aggregate the total building demand
head(aggregate_timeseries(df, varname = "total_buildings"))

}
