% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{approximate_sessions}
\alias{approximate_sessions}
\title{Approximate sessions to perfect power steps}
\usage{
approximate_sessions(sessions, time_interval = 15, power_interval = 1)
}
\arguments{
\item{sessions}{sessions data set}

\item{time_interval}{interval of time approximation, in minutes}

\item{power_interval}{interval of power approximation, in kW}
}
\value{
tibble
}
\description{
Approximate sessions to perfect power steps
}
