% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battery_optimization.R
\name{battery_solve_grid_window}
\alias{battery_solve_grid_window}
\title{Perform battery optimization for a single window (grid/capacity objective)}
\usage{
battery_solve_grid_window(
  G,
  L,
  Bcap,
  Bc,
  Bd,
  SOCmin,
  SOCmax,
  SOCini,
  import_capacity,
  export_capacity,
  P,
  q
)
}
\arguments{
\item{G}{numeric vector, renewable generation profile}

\item{L}{numeric vector, static load profile}

\item{Bcap}{numeric, battery capacity in energy units (kWh * slots/h)}

\item{Bc}{numeric, maximum charging power (kW)}

\item{Bd}{numeric, maximum discharging power (kW)}

\item{SOCmin}{numeric, minimum State-of-Charge (\%)}

\item{SOCmax}{numeric, maximum State-of-Charge (\%)}

\item{SOCini}{numeric, initial State-of-Charge (\%)}

\item{import_capacity}{numeric vector, maximum grid import (kW)}

\item{export_capacity}{numeric vector, maximum grid export (kW)}

\item{P}{numeric matrix, quadratic objective term}

\item{q}{numeric vector, linear objective term}
}
\value{
numeric vector
}
\description{
Perform battery optimization for a single window (grid/capacity objective)
}
\keyword{internal}
