% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{change_timeseries_resolution}
\alias{change_timeseries_resolution}
\title{Change time resolution of a time-series data frame}
\usage{
change_timeseries_resolution(df, resolution_out, method)
}
\arguments{
\item{df}{tibble, first column of name \code{datetime} being of class datetime and rest of columns being numeric}

\item{resolution_out}{integer, desired interval of minutes between two consecutive datetime values}

\item{method}{character, being \code{interpolate}, \code{repeat} or \code{divide} if the resolution has to be increased,
or \code{average}, \code{first} or \code{sum} if the resolution has to be decreased. See Examples for more information.}
}
\value{
tibble
}
\description{
Change time resolution of a time-series data frame
}
\examples{
# Example time-series data set: time resolution 15 minutes
head(energy_profiles)

# Change time resolution to 60 minutes
# It is decreasing time resolution, so we use the `average` method
head(change_timeseries_resolution(
  df = energy_profiles,
  resolution_out = 60,
  method = "average"
))



}
