% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{convert_datetime_to_timeslot}
\alias{convert_datetime_to_timeslot}
\title{Convert datetime to time slot index}
\usage{
convert_datetime_to_timeslot(dttm, start, time_interval_mins)
}
\arguments{
\item{dttm}{datetime, value to convert to time slot index}

\item{start}{datetime, start datetime value}

\item{time_interval_mins}{integer, interval of time between time slots (minutes)}
}
\description{
Convert datetime to time slot index
}
