% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{convert_timeslot_to_datetime}
\alias{convert_timeslot_to_datetime}
\title{Convert time slot index to datetime}
\usage{
convert_timeslot_to_datetime(timeslot, start, time_interval)
}
\arguments{
\item{timeslot}{integer, time slot}

\item{start}{datetime, start datetime value}

\item{time_interval}{integer, interval of time between time slots (minutes)}
}
\description{
Convert time slot index to datetime
}
