% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{curtail_capacity_window_battery}
\alias{curtail_capacity_window_battery}
\title{Battery optimal charging/discharging profile to minimize grid interaction (just a window)}
\usage{
curtail_capacity_window_battery(
  G,
  L,
  Bcap,
  Bc,
  Bd,
  SOCmin,
  SOCmax,
  SOCini,
  import_capacity,
  export_capacity,
  lambda = 0
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{Bcap}{numeric, capacity of the battery (NOT in kWh but in energy units according to time resolution)}

\item{Bc}{numeric, maximum charging power (in kW)}

\item{Bd}{numeric, maximum discharging power (in kW)}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{import_capacity}{numeric or numeric vector, grid maximum import power capacity that will limit the maximum charging power}

\item{export_capacity}{numeric or numeric vector, grid maximum export power capacity that will limit the maximum discharging power}

\item{lambda}{numeric, penalty on change for the flexible load.}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile to minimize grid interaction (just a window)
}
\keyword{internal}
