% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{denormalize_sessions}
\alias{denormalize_sessions}
\title{Denormalize sessions (from timeslot to datetime)}
\usage{
denormalize_sessions(sessions_norm, start, time_interval)
}
\arguments{
\item{sessions_norm}{normalized sessions data set}

\item{start}{first value of the normalization datetime sequence}

\item{time_interval}{interval of time (minutes)}
}
\value{
tibble
}
\description{
Denormalize sessions (from timeslot to datetime)
}
