% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{denormalize_timeseries}
\alias{denormalize_timeseries}
\title{Denormalize time-series (from timeslot to demand)}
\usage{
denormalize_timeseries(df, start, time_interval)
}
\arguments{
\item{df}{tibble or data.frame with first column being \code{timeslot}}

\item{start}{datetime, start datetime value}

\item{time_interval}{integer, interval of time between time slots (minutes)}
}
\value{
tibble with first column being \code{datetime}
}
\description{
Denormalize time-series (from timeslot to demand)
}
