% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energy_profiles}
\alias{energy_profiles}
\title{Time-series profiles for consumption, production and energy prices}
\format{
A tibble
\describe{
\item{solar}{Solar PV production (in kW)}
\item{building}{Consumption profile of a building (in kW)}
\item{price_imported}{Imported energy price (in €/kWh)}
\item{price_exported}{Exported energy price (in €/kWh)}
\item{price_turn_up}{Balancing price for increasing demand (in €/kWh)}
\item{price_turn_down}{Balancing price for decreasing demand (in €/kWh)}
}
}
\usage{
energy_profiles
}
\description{
A tibble with time-series profiles for solar production, building consumption
and energy prices
}
\keyword{internal}
