% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpower.R
\name{expand_session}
\alias{expand_session}
\title{Divide a session into multiple sessions according to time resolution}
\usage{
expand_session(session, resolution)
}
\arguments{
\item{session}{tibble, charging sessions data set}

\item{resolution}{integer, time resolution in minutes}
}
\value{
tibble
}
\description{
Divide a session into multiple sessions according to time resolution
}
