% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-evsim.R
\name{expand_sessions}
\alias{expand_sessions}
\title{#' @importFrom evsim get_demand
#' @export
evsim::get_demand}
\usage{
expand_sessions(sessions, resolution)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{evprof} package
(see \href{https://mcanigueral.github.io/evprof/articles/sessions-format.html}{this article})}

\item{resolution}{integer, time resolution (in minutes) of the time slots}
}
\value{
tibble
}
\description{
#' @importFrom evsim expand_sessions
#' @export
evsim::expand_sessions
}
\details{
Expand sessions along time slots

Every session in \code{sessions} is divided in multiple time slots
with the corresponding \code{Power} consumption, among other variables.

The \code{Power} value is calculated for every time slot according to the original
required energy. The columns \code{NominalPower}, \code{RequiredEnergy} and
\code{FlexibilityHours} correspond to the values of the original session, and not
to the expanded session in every time slot. The column \code{ID} shows the number
of the time slot corresponding to the original session.
}
\keyword{interal}
