% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_conversion_losses}
\alias{get_conversion_losses}
\title{Losses due to charging/discharging process}
\usage{
get_conversion_losses(power, charge_eff = 1, discharge_eff = 1)
}
\arguments{
\item{power}{numeric vector, being positive when charging and negative when discharging}

\item{charge_eff}{numeric, charging efficiency (from 0 to 1, default to 1)}

\item{discharge_eff}{numeric, discharging efficiency (from 0 to 1, default to 1)}
}
\value{
numeric vector
}
\description{
Losses due to charging/discharging process
}
