% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_demand}
\alias{get_demand}
\title{Get demand}
\usage{
get_demand(
  sessions,
  dttm_seq,
  aggregated = FALSE,
  stacked = FALSE,
  normalized = FALSE
)
}
\arguments{
\item{sessions}{sessions data set}

\item{dttm_seq}{vector with a sequence of datetime values}

\item{aggregated}{single "Demand" column or one column for each profile}

\item{stacked}{single "Profile" column or one column for each profile}

\item{normalized}{True if sessions datatset is normalized (timeslots instead of datetimes)}
}
\value{
data.frame
}
\description{
Get demand
}
