% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_dhm_idx}
\alias{get_dhm_idx}
\title{Get initial index}
\usage{
get_dhm_idx(dt, time_interval_mins = 15)
}
\arguments{
\item{dt}{datetime}

\item{time_interval_mins}{interval of time (minutes)}
}
\description{
Get initial index
}
