% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_duration_curve}
\alias{get_duration_curve}
\title{Duration curve table for a given vector}
\usage{
get_duration_curve(vct)
}
\arguments{
\item{vct}{numeric vector}
}
\value{
tibble
}
\description{
The Duration Curve (DC) represents the percentage of time that a specific
value of a vector has been used in a specific time period.
}
\keyword{internal}
