% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_energy}
\alias{get_energy}
\title{Get energy from charging hours and charging power}
\usage{
get_energy(power_kW, charging_start, charging_end, time_interval_mins)
}
\arguments{
\item{power_kW}{charging power (kW)}

\item{charging_start}{connection start hour}

\item{charging_end}{charging end hour}

\item{time_interval_mins}{interval of time (minutes)}
}
\description{
Get energy from charging hours and charging power
}
