% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_energy_balance}
\alias{get_energy_balance}
\title{Get energy balance time-series}
\usage{
get_energy_balance(df)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production}}
}
\value{
tibble
}
\description{
Input data frame must have columns \code{consumption}, \code{production}.
Output data frame has extra columns \code{net}, \code{local}, \code{imported}, \code{exported}.
Column \code{net} is positive when there is more consumption than production
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  production = solar,
  consumption = building
)
head(df)

head(get_energy_balance(df))


}
