% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_energy_cost}
\alias{get_energy_cost}
\title{Get energy cost time-series}
\usage{
get_energy_cost(df)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production}, \code{price_imported} and \code{price_exported}}
}
\value{
tibble
}
\description{
The energy cost (in Euros) based on imported and exported energy
and the corresponding prices.
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  production = solar,
  consumption = building,
  price_imported,
  price_exported
)
df <- dplyr::slice_head(df, n = 300)
head(df)

head(get_energy_cost(df))

}
