% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_energy_total_cost}
\alias{get_energy_total_cost}
\title{Get energy total cost (one number)}
\usage{
get_energy_total_cost(df)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production}, \code{price_imported} and \code{price_exported}}
}
\value{
tibble
}
\description{
The total energy cost (in Euros) based on imported and exported energy
and the corresponding prices.
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  production = solar,
  consumption = building,
  price_imported,
  price_exported
)
head(df)

get_energy_total_cost(df)

}
