% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_imbalance_income}
\alias{get_imbalance_income}
\title{Get imbalance income}
\usage{
get_imbalance_income(df)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{demand_baseline}, \code{demand_final}, \code{price_turn_up} and \code{price_turn_down}}
}
\value{
tibble
}
\description{
The energy income (in Euros) based on the difference between the baseline and
final power profile for a certain demand. The difference represents the
optimal changes due to flexibility, so this function calculates the income
from applying this demand-side flexibility.
}
