% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_imbalance_total_income}
\alias{get_imbalance_total_income}
\title{Get imbalance total income (one number)}
\usage{
get_imbalance_total_income(df)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{demand_baseline}, \code{demand_final}, \code{price_turn_up} and \code{price_turn_down}}
}
\value{
tibble
}
\description{
The total energy income (in Euros) based on the difference between the baseline and
final power profile for a certain demand. The difference represents the
optimal changes due to flexibility, so this function calculates the income
from applying this demand-side flexibility.
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  demand_baseline = building,
  price_turn_up,
  price_turn_down
)

# Build another random consumption profile
building_variation <- rnorm(nrow(df), mean = 0, sd = 1)
df <- dplyr::mutate(
  df, demand_final = demand_baseline + building_variation
)
head(df)

get_imbalance_total_income(df)

}
