% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_load_duration_curve}
\alias{get_load_duration_curve}
\title{Load duration curve table}
\usage{
get_load_duration_curve(df)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production}}
}
\value{
tibble
}
\description{
The Load Duration Curve (LDC) represents the percentage of time that a specific
value of power has been used in the electrical grid for a specific time period.
It's widely used for power system planning and grid reliability assessments.
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  consumption = building
)

head(df)

get_load_duration_curve(df)

}
