% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_percentage_hours}
\alias{get_percentage_hours}
\title{Get percentage of hours for load duration curve}
\usage{
get_percentage_hours(vct, threshold)
}
\arguments{
\item{vct}{numeric vector of power}

\item{threshold}{numeric, threshold of power}
}
\value{
numeric
}
\description{
The percentage of hours that a certain threshold of power is surpassed
}
\keyword{internal}
