% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_schedule}
\alias{get_schedule}
\title{Get schedule}
\usage{
get_schedule(sessions_norm, window)
}
\arguments{
\item{sessions_norm}{dataframe of normalized sessions}

\item{window}{vector with time window's indices}
}
\value{
data.frame
}
\description{
Get schedule
}
