% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_sessions_demand}
\alias{get_sessions_demand}
\title{Obtain timeseries demand from sessions dataset}
\usage{
get_sessions_demand(
  sessions,
  dttm_seq = NULL,
  by = "Profile",
  normalized = F,
  resolution = 15
)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{dttm_seq}{sequence of datetime values that will be the datetime variable of the returned time-series data frame.
This can be an integer sequence as well if \code{normalized = TRUE}.}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}

\item{normalized}{logical, whether the \code{sessions} datetime columns are time-slot values}

\item{resolution}{integer, time resolution (in minutes) of the sessions datetime variables. If \code{dttm_seq} is defined this parameter is ignored.}
}
\value{
tibble
}
\description{
Obtain timeseries demand from sessions dataset
}
