% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_sessions_interval_demand}
\alias{get_sessions_interval_demand}
\title{Obtain demand from a starting dttm value and certain duration interval}
\usage{
get_sessions_interval_demand(sessions, timeslot, by, normalized)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{timeslot}{datetime, time slot of the requested demand}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}

\item{normalized}{logical, whether the \code{sessions} datetime columns are time-slot values}
}
\value{
tibble
}
\description{
Obtain demand from a starting dttm value and certain duration interval
}
