% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2g.R
\name{get_setpoints_v2g}
\alias{get_setpoints_v2g}
\title{Set setpoints for smart V2G}
\usage{
get_setpoints_v2g(
  sessions_window,
  opt_data,
  profiles_demand,
  opt_objective,
  lambda
)
}
\description{
Set setpoints for smart V2G
}
\keyword{internal}
