% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpower.R
\name{limit_charging_power}
\alias{limit_charging_power}
\title{Simulate charging power limitation based on grid capacity signals}
\usage{
limit_charging_power(
  sessions,
  grid_capacity,
  amps_car_min = 8,
  expand = TRUE,
  include_log = FALSE
)
}
\arguments{
\item{sessions}{tibble, charging sessions data set.
If sessions have been expanded using \code{expand_session} function, set parameter \code{expand = FALSE}.}

\item{grid_capacity}{tibble, grid capacity time-series with columns \code{datetime} and \code{max_amps}}

\item{amps_car_min}{integer, minimum current to charge the vehicles}

\item{expand}{logical, whether to expand session using \code{expand_session} function}

\item{include_log}{logical, whether to print and return log messages}
}
\value{
tibble
}
\description{
Simulate charging power limitation based on grid capacity signals
}
