% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minimize_cost}
\alias{minimize_cost}
\title{Minimization of energy cost}
\usage{
minimize_cost(
  G,
  LF,
  LS = NULL,
  PI,
  PE,
  direction = "forward",
  time_horizon = NULL,
  window_length = NULL,
  flex_window_length = window_length,
  LFmax = Inf,
  grid_capacity = Inf,
  mc.cores = 2
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{LF}{numeric vector, being the flexible load profile}

\item{LS}{numeric vector, being the static load profile}

\item{PI}{numeric vector, electricity prices for imported energy}

\item{PE}{numeric vector, electricity prices for exported energy}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from}

\item{window_length}{integer, window length in time slots (not hours).
If \code{NULL}, the window length will be the length of \code{G}, \code{LF}, and \code{LS} (it must be the same)
It can also be an integer vector, specifying a specific length for every
window along the length of \code{G}. In that case, the sum of all vector values
must be equal to the length of \code{G}.}

\item{flex_window_length}{integer, flexibility window length in time slots (not hours).
This optional feature lets you apply flexibility only during few hours from the start of the window.}

\item{LFmax}{numeric, value of maximum power (in kW) of the flexible load \code{LF}}

\item{grid_capacity}{numeric or numeric vector, grid maximum power capacity that will limit the maximum optimized demand}

\item{mc.cores}{integer, number of cores to use.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
numeric vector
}
\description{
Minimization of energy cost
}
