% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minimize_cost_window}
\alias{minimize_cost_window}
\title{Minimization of the cost (just a window)}
\usage{
minimize_cost_window(
  G,
  LF,
  LS,
  PI,
  PE,
  PTD,
  PTU,
  direction,
  time_horizon,
  LFmax,
  import_capacity,
  export_capacity,
  lambda = 0
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation power profile}

\item{LF}{numeric vector, being the flexible load power profile}

\item{LS}{numeric vector, being the static load power profile}

\item{PI}{numeric vector, electricity prices for imported energy}

\item{PE}{numeric vector, electricity prices for exported energy}

\item{PTD}{numeric vector, prices for turn-down energy use}

\item{PTU}{numeric vector, prices for turn-up energy use}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from}

\item{LFmax}{numeric, value of maximum power (in kW) of the flexible load \code{LF}}

\item{import_capacity}{numeric or numeric vector, grid maximum import capacity that will limit the maximum optimized demand}

\item{export_capacity}{numeric or numeric vector, grid maximum export capacity that will limit the maximum optimized demand}

\item{lambda}{numeric, penalty on change for the flexible load.}
}
\value{
numeric vector
}
\description{
Minimization of the cost (just a window)
}
\keyword{internal}
