% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minimize_cost_window_battery}
\alias{minimize_cost_window_battery}
\title{Battery optimal charging/discharging profile to minimize cost (just a window)}
\usage{
minimize_cost_window_battery(
  G,
  L,
  PE,
  PI,
  PTD,
  PTU,
  Bcap,
  Bc,
  Bd,
  SOCmin,
  SOCmax,
  SOCini,
  import_capacity,
  export_capacity,
  lambda = 0
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{PE}{numeric vector, electricity prices for exported energy}

\item{PI}{numeric vector, electricity prices for imported energy}

\item{PTD}{numeric vector, prices for turn-down energy use}

\item{PTU}{numeric vector, prices for turn-up energy use}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{import_capacity}{numeric or numeric vector, grid maximum import power capacity that will limit the maximum charging power}

\item{export_capacity}{numeric or numeric vector, grid maximum export power capacity that will limit the maximum discharging power}

\item{lambda}{numeric, penalty on change for the flexible load.}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile to minimize cost (just a window)
}
\keyword{internal}
