% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minimize_grid_flow_window}
\alias{minimize_grid_flow_window}
\title{Minimization of the grid flow (just a window)}
\usage{
minimize_grid_flow_window(
  G,
  LF,
  LS,
  direction,
  time_horizon,
  LFmax,
  grid_capacity,
  lambda = 0
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{LF}{numeric vector, being the flexible load profile}

\item{LS}{numeric vector, being the static load profile}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from}

\item{LFmax}{numeric, value of maximum power (in kW) of the flexible load \code{LF}}

\item{grid_capacity}{numeric or numeric vector, grid maximum power capacity that will limit the maximum optimized demand}

\item{lambda}{numeric, penalty on change for the flexible load.
This is a factor used in the optimization problem.}
}
\value{
numeric vector
}
\description{
Minimization of the grid flow (just a window)
}
\keyword{internal}
