% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minimize_grid_flow_window_osqp}
\alias{minimize_grid_flow_window_osqp}
\title{Minimization of the grid flow (just a window)}
\usage{
minimize_grid_flow_window_osqp(
  w,
  G,
  LF,
  LS = NULL,
  direction = "forward",
  time_horizon = NULL,
  up_to_G = TRUE
)
}
\arguments{
\item{w}{numeric between \code{0} and \code{1}, being the weight of the flexibility potential}

\item{G}{numeric vector, being the renewable generation profile}

\item{LF}{numeric vector, being the flexible load profile}

\item{LS}{numeric vector, being the static load profile}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from}

\item{up_to_G}{logical, whether to limit the flexible EV demand up to renewable Generation}
}
\value{
numeric vector
}
\description{
Minimization of the grid flow (just a window)
}
