% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2g.R
\name{minimize_net_power_v2g_window}
\alias{minimize_net_power_v2g_window}
\title{Minimisation of net power with bidirectional flexibility (single window)}
\usage{
minimize_net_power_v2g_window(
  G,
  LF,
  LS,
  direction,
  time_horizon,
  LFmax,
  import_capacity,
  export_capacity,
  lambda = 0
)
}
\description{
Minimisation of net power with bidirectional flexibility (single window)
}
\keyword{internal}
