% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minimize_net_power_window_battery}
\alias{minimize_net_power_window_battery}
\title{Battery optimal charging/discharging profile to minimize grid interaction (just a window)}
\usage{
minimize_net_power_window_battery(
  G,
  L,
  Bcap,
  Bc,
  Bd,
  SOCmin,
  SOCmax,
  SOCini,
  grid_capacity,
  lambda = 0
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{grid_capacity}{numeric or numeric vector, grid maximum power capacity that will limit the maximum optimized demand}

\item{lambda}{numeric, penalty on change for the battery power profile.
This is a factor used in the optimization problem.}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile to minimize grid interaction (just a window)
}
\keyword{internal}
