% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalize_sessions}
\alias{normalize_sessions}
\title{Normalize sessions (from datetime to timeslots)}
\usage{
normalize_sessions(sessions, start, time_interval)
}
\arguments{
\item{sessions}{sessions data set}

\item{start}{first value of the normalization datetime sequence}

\item{time_interval}{interval of time (minutes)}
}
\description{
Normalize sessions (from datetime to timeslots)
}
