% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimize_battery_window}
\alias{optimize_battery_window}
\title{Battery optimal charging/discharging profile to minimize net power and cost (just a window)}
\usage{
optimize_battery_window(
  G,
  L,
  PE,
  PI,
  PTD,
  PTU,
  Bcap,
  Bc,
  Bd,
  SOCmin,
  SOCmax,
  SOCini,
  grid_capacity,
  w,
  lambda
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{PE}{numeric vector, electricity prices for exported energy}

\item{PI}{numeric vector, electricity prices for imported energy}

\item{PTD}{numeric vector, prices for turn-down energy use}

\item{PTU}{numeric vector, prices for turn-up energy use}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{grid_capacity}{numeric or numeric vector, grid maximum power capacity that will limit the maximum optimized demand}

\item{w}{numeric, optimization objective weight (\code{w=1} minimizes net power while \code{w=0} minimizes cost).}

\item{lambda}{numeric, penalty on change for the battery power profile.
This is a factor used in the optimization problem.}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile to minimize net power and cost (just a window)
}
\keyword{internal}
