% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-charging.R
\name{plot.SmartCharging}
\alias{plot.SmartCharging}
\title{\code{plot} method for \code{SmartCharging} object class}
\usage{
\method{plot}{SmartCharging}(x, ...)
}
\arguments{
\item{x}{\code{SmartCharging} object returned by \code{smart_charging}  function}

\item{...}{further arguments passed to or from other methods.}
}
\value{
HTML interctive plot
}
\description{
\code{plot} method for \code{SmartCharging} object class
}
\keyword{internal}
