% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{plot_energy_cost}
\alias{plot_energy_cost}
\title{Plot hourly energy cost}
\usage{
plot_energy_cost(df, original_df = NULL, ...)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production},
\code{price_imported} and \code{price_exported}. If \code{imbalance = TRUE}, then columns
\code{demand_baseline}, \code{demand_final}, \code{price_turn_up} and \code{price_turn_down} are
also required.}

\item{original_df}{tibble with same columns than \code{df} corresponding to the
original scenario (e.g. without flexibility).}

\item{...}{extra arguments to pass to dygraphs::dyOptions function.}
}
\value{
dygraphs plot
}
\description{
HTML interactive plot showing the hourly energy cost. Also,
a comparison between the original scenario is done when
\code{original_df} is not \code{NULL}.
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  production = solar,
  consumption = building,
  price_imported,
  price_exported
)
df <- dplyr::slice_head(df, n = 300)
head(df)

plot_energy_cost(df)

# Build another random building profile
building_variation <- rnorm(nrow(df), mean = 0, sd = 1)
df2 <- dplyr::mutate(df, consumption = consumption + building_variation)
plot_energy_cost(df2, original_df = df)

}
