% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{plot_load_duration_curve}
\alias{plot_load_duration_curve}
\title{Plot of the load duration curve}
\usage{
plot_load_duration_curve(df, original_df = NULL)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production}}

\item{original_df}{tibble with same columns than \code{df} corresponding to the
original scenario (e.g. without flexibility).}
}
\value{
dygraphs plot
}
\description{
HTML interactive plot showing the graphical representation of a load duration curve.
The Load Duration Curve (LDC) represents the percentage of time that a specific
value of power has been used in the electrical grid for a specific time period.
It's widely used for power system planning and grid reliability assessments.
Also, a comparison between the original scenario is done when
\code{original_df} is not \code{NULL}.
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  consumption = building
)
head(df)

plot_load_duration_curve(df)

# Build another random building profile
building_variation <- rnorm(nrow(df), mean = 0, sd = 1)
df2 <- dplyr::mutate(df, consumption = consumption + building_variation)
plot_load_duration_curve(df2, original_df = df)

}
