% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-charging.R
\name{print.SmartCharging}
\alias{print.SmartCharging}
\title{\code{print} method for \code{SmartCharging} object class}
\usage{
\method{print}{SmartCharging}(x, ...)
}
\arguments{
\item{x}{\code{SmartCharging} object returned by \code{smart_charging}  function}

\item{...}{further arguments passed to or from other methods.}
}
\value{
nothing but prints information about the \code{SmartCharging} object
}
\description{
\code{print} method for \code{SmartCharging} object class
}
\keyword{internal}
