% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-evsim.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{get_demand}
\alias{expand_sessions}
\alias{adapt_charging_features}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{evsim}{\code{\link[evsim]{adapt_charging_features}}, \code{\link[evsim]{expand_sessions}}, \code{\link[evsim]{get_demand}}}
}}

