% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-charging.R
\name{smart_charging_window}
\alias{smart_charging_window}
\title{Set setpoints for smart charging}
\usage{
smart_charging_window(
  sessions_window,
  profiles_demand,
  setpoints,
  method,
  power_th = 0,
  charging_power_min = 0,
  energy_min = 1,
  energy_max = 1,
  include_log = FALSE
)
}
\arguments{
\item{sessions_window}{tibble, sessions corresponding to a single windows}

\item{profiles_demand}{tibble, user profiles power demand}

\item{setpoints}{tibble, user profiles power setpoints}

\item{method}{character, scheduling method being \code{"none"}, \code{"postpone"}, \code{"curtail"} or \code{"interrupt"}.
If \code{none}, the scheduling part is skipped and the sessions returned in the
results will be identical to the original parameter.}

\item{power_th}{numeric, power threshold (between 0 and 1) accepted from setpoint.
For example, with \code{power_th = 0.1} and \code{setpoint = 100} for a certain time slot,
then sessions' demand can reach a value of \code{110} without needing to schedule sessions.}

\item{charging_power_min}{numeric. It can be configured in two ways:
(1) minimum allowed ratio (between 0 and 1) of nominal power (i.e. \code{Power} column in \code{sessions}), or
(2) specific value of minimum power (in kW) higher than 1 kW.

For example, if \code{charging_power_min = 0.5} and \code{method = 'curtail'}, sessions' charging power can only
be curtailed until the 50\% of the nominal charging power.
And if \code{charging_power_min = 2}, sessions' charging power can be curtailed until 2 kW.}

\item{energy_min}{numeric, minimum allowed ratio (between 0 and 1) of required energy.}

\item{energy_max}{numeric, maximum allowed ratio (between 0 and 1) of required energy.}

\item{include_log}{logical, whether to output the algorithm messages for every user profile and time-slot}
}
\description{
Set setpoints for smart charging
}
\keyword{internal}
