% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{solve_optimization_battery_window}
\alias{solve_optimization_battery_window}
\title{Perform battery optimization (just a window)}
\usage{
solve_optimization_battery_window(
  P,
  q,
  G,
  L,
  Bcap,
  Bc,
  Bd,
  SOCmin,
  SOCmax,
  SOCini,
  import_capacity,
  export_capacity,
  charge_eff,
  discharge_eff
)
}
\arguments{
\item{P}{numeric matrix, optimization objective parameter}

\item{q}{numeric vector, optimization objective parameter}

\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{Bcap}{numeric, capacity of the battery (NOT in kWh but in energy units according to time resolution)}

\item{Bc}{numeric, maximum charging power (in kW)}

\item{Bd}{numeric, maximum discharging power (in kW)}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{import_capacity}{numeric or numeric vector, grid maximum import power capacity that will limit the maximum charging power}

\item{export_capacity}{numeric or numeric vector, grid maximum export power capacity that will limit the maximum discharging power}

\item{charge_eff}{numeric, battery charging efficiency (default 1).}

\item{discharge_eff}{numeric, battery discharging efficiency (default 1).}
}
\value{
numeric vector
}
\description{
Perform battery optimization (just a window)
}
\keyword{internal}
