% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{solve_optimization_window}
\alias{solve_optimization_window}
\title{Perform optimization (just a window)}
\usage{
solve_optimization_window(
  G,
  LF,
  LS,
  direction,
  time_horizon,
  LFmax,
  import_capacity,
  export_capacity,
  P,
  q
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation power profile}

\item{LF}{numeric vector, being the flexible load power profile}

\item{LS}{numeric vector, being the static load power profile}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from}

\item{LFmax}{numeric, value of maximum power (in kW) of the flexible load \code{LF}}

\item{import_capacity}{numeric or numeric vector, grid maximum import capacity that will limit the maximum optimized demand}

\item{export_capacity}{numeric or numeric vector, grid maximum export capacity that will limit the maximum optimized demand}
}
\value{
numeric vector
}
\description{
Perform optimization (just a window)
}
\keyword{internal}
