% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-charging.R
\name{summarise_by_segment}
\alias{summarise_by_segment}
\title{Summarise sessions by segment}
\usage{
summarise_by_segment(ss)
}
\arguments{
\item{ss}{sessions expanded schedule from \code{schedule_sessions} function}
}
\description{
Simplify the extended sessions schedule by joining consecutive rows with
same charging power (power segments).
}
\keyword{internal}
