% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{summarise_energy_charged}
\alias{summarise_energy_charged}
\title{Get a summary of the energy charged per session}
\usage{
summarise_energy_charged(smart_charging, sessions)
}
\arguments{
\item{smart_charging}{SmartCharging object, returned by function \code{smart_charging()}}

\item{sessions}{tibble, sessions data set containig the following variables:
\code{"Session"}, \code{"Timecycle"}, \code{"Profile"}, \code{"ConnectionStartDateTime"}, \code{"ConnectionHours"}, \code{"Power"} and \code{"Energy"}.}
}
\value{
tibble
}
\description{
A table is provided containing the originally required energy for every
session and the actual energy charged with the smart charging program
(also in percentage).
}
\examples{
library(dplyr)

sessions <- evsim::california_ev_sessions_profiles \%>\%
  slice_head(n = 50) \%>\%
  evsim::adapt_charging_features(time_resolution = 15)
sessions_demand <- evsim::get_demand(sessions, resolution = 15)

# Don't require any other variable than datetime, since we don't
# care about local generation (just peak shaving objective)
opt_data <- tibble(
  datetime = sessions_demand$datetime,
  production = 0
)
sc_results <- smart_charging(
  sessions, opt_data, opt_objective = "grid", method = "curtail",
  window_days = 1, window_start_hour = 6,
  responsive = list(Workday = list(Worktime = 0.9)),
  energy_min = 0.5
)

summarise_energy_charged(sc_results, sessions)

}
