% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-charging.R
\name{summarise_smart_charging_sessions}
\alias{summarise_smart_charging_sessions}
\title{Get a summary of the new schedule of charging sessions}
\usage{
summarise_smart_charging_sessions(smart_charging)
}
\arguments{
\item{smart_charging}{SmartCharging object, returned by function \code{smart_charging()}}
}
\value{
tibble with columns:
\code{timecycle} (time-cycle name),
\code{profile} (user profile name),
\code{group} (name of sessions' group),
\code{subgroup} (nome of sessions' subgroup),
\code{n_sessions} (number of sessions) and
\code{pct} (percentage of subgroup sessions from the group)
}
\description{
A table is provided containing the number of \code{Considered}, \code{Responsive},
\code{Flexbile} and \code{Exploited} sessions, by user profile.
}
\examples{
library(dplyr)

# Use first 50 sessions
sessions <- evsim::california_ev_sessions_profiles \%>\%
  slice_head(n = 50) \%>\%
  evsim::adapt_charging_features(time_resolution = 15)
sessions_demand <- evsim::get_demand(sessions, resolution = 15)

# Don't require any other variable than datetime, since we don't
# care about local generation (just peak shaving objective)
opt_data <- tibble(
  datetime = sessions_demand$datetime,
  production = 0
)
sc_results <- smart_charging(
  sessions, opt_data,
  opt_objective = "grid", method = "curtail",
  window_days = 1, window_start_hour = 6,
  responsive = list(Workday = list(Worktime = 0.9)),
  energy_min = 0.5
)

summarise_smart_charging_sessions(sc_results)

}
