% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{adapt_dttm_seq_to_opt_windows}
\alias{adapt_dttm_seq_to_opt_windows}
\title{Cut the dateteime sequence according to the optimizion window start hour and length}
\usage{
adapt_dttm_seq_to_opt_windows(dttm_seq, window_start_hour)
}
\arguments{
\item{dttm_seq}{datetime sequence of the original timeseries data}

\item{window_start_hour}{integer, hour to start the optimization window}
}
\value{
datetime sequence
}
\description{
Cut the dateteime sequence according to the optimizion window start hour and length
}
