% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battery_optimization.R
\name{add_battery_optimization}
\alias{add_battery_optimization}
\title{Battery optimal charging/discharging profile}
\usage{
add_battery_optimization(
  opt_data,
  opt_objective = "grid",
  Bcap,
  Bc,
  Bd,
  SOCmin = 0,
  SOCmax = 100,
  SOCini = NULL,
  window_days = 1,
  window_start_hour = 0,
  flex_window_hours = 24,
  lambda = 0,
  charge_eff = 1,
  discharge_eff = 1
)
}
\arguments{
\item{opt_data}{tibble with optimization data}

\item{opt_objective}{optimization objective}

\item{Bcap}{numeric, battery capacity in kWh}

\item{Bc}{numeric, max charging power}

\item{Bd}{numeric, max discharging power}

\item{SOCmin}{numeric, minimum state of charge in \%}

\item{SOCmax}{numeric, maximum state of charge in \%}

\item{SOCini}{numeric, initial state of charge in \%}

\item{window_days}{integer, optimization window size in days}

\item{window_start_hour}{integer, hour when each window starts}

\item{flex_window_hours}{integer, active optimization hours per window}

\item{lambda}{numeric, smoothing penalty}

\item{charge_eff}{numeric, charging efficiency}

\item{discharge_eff}{numeric, discharging efficiency}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile
}
