% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{add_battery_optimization}
\alias{add_battery_optimization}
\title{Battery optimal charging/discharging profile}
\usage{
add_battery_optimization(
  w,
  G,
  L,
  Bcap,
  Bc,
  Bd,
  SOCmin = 0,
  SOCmax = 100,
  SOCini = NULL,
  window_length = NULL
)
}
\arguments{
\item{w}{numeric between \code{0} and \code{1}, being the weight of the flexibility potential}

\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{window_length}{integer, window length in time slots (not hours).
If \code{NULL}, the window length will be the length of \code{G}.
It can also be an integer vector, specifying a specific length for every
window along the length of \code{G}. In that case, the sum of all vector values
must be equal to the length of \code{G}.}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile
}
