% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{add_battery_optimization}
\alias{add_battery_optimization}
\title{Battery optimal charging/discharging profile}
\usage{
add_battery_optimization(
  opt_data,
  opt_objective = "grid",
  Bcap,
  Bc,
  Bd,
  SOCmin = 0,
  SOCmax = 100,
  SOCini = NULL,
  window_days = 1,
  window_start_hour = 0,
  flex_window_hours = 24,
  lambda = 0,
  mc.cores = 1
)
}
\arguments{
\item{opt_data}{tibble, optimization contextual data.
The first column must be named \code{datetime} (mandatory) containing the
date time sequence where the optimization algorithm is applied.
The other columns can be (optional):
\itemize{
\item \code{static}: static power demand (in kW) from other sectors like buildings,
offices, etc.
\item \code{import_capacity}: maximum imported power from the grid (in kW),
for example the contracted power with the energy company.
\item \code{export_capacity}: maximum exported power from the grid (in kW),
for example the contracted power with the energy company.
\item \code{production}: local power generation (in kW).
This is used when \code{opt_objective = "grid"}.
\item \code{price_imported}: price for imported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_exported}: price for exported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_turn_down}: price for turn-down energy use (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_turn_up}: price for turn-up energy use (€/kWh).
This is used when \code{opt_objective = "cost"}.
}}

\item{opt_objective}{character or numeric.
Optimization objective can be \code{"grid"} (default), \code{"cost"} or \code{"curtail"}, or
a number between \code{0} and \code{1} to perform combined optimization
where \code{0 == "cost"} and \code{1 == "grid"}.}

\item{Bcap}{numeric, capacity of the battery (in kWh)}

\item{Bc}{numeric, maximum charging power (in kW)}

\item{Bd}{numeric, maximum discharging power (in kW)}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{window_days}{integer, number of days to consider as optimization window.}

\item{window_start_hour}{integer, starting hour of the optimization window.}

\item{flex_window_hours}{integer, flexibility window length, in hours.
This optional feature lets you apply flexibility only during few hours from the \code{window_start_hour}.
It must be lower than \code{window_days*24} hours.}

\item{lambda}{numeric, penalty on change for the flexible load.}

\item{mc.cores}{integer, number of cores to use.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
numeric vector
}
\description{
See the formulation of the optimization problems in the
\href{https://resourcefully-dev.github.io/flextools/}{documentation website}.
}
\examples{
library(dplyr)
opt_data <- flextools::energy_profiles \%>\%
  filter(lubridate::isoweek(datetime) == 18) \%>\%
  rename(
    production = "solar"
  ) \%>\%
  select(any_of(c(
    "datetime", "production", "building", "price_imported", "price_exported"
  ))) \%>\%
  mutate(
    static = .data$building
  )
  opt_battery <- opt_data \%>\%
    add_battery_optimization(
      opt_objective = 0.5,
      Bcap = 50, Bc = 4, Bd = 4,
      window_start_hour = 5
    )

}
