% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{add_battery_optimization}
\alias{add_battery_optimization}
\title{Battery optimal charging/discharging profile}
\usage{
add_battery_optimization(
  opt_data,
  opt_objective = "grid",
  Bcap,
  Bc,
  Bd,
  SOCmin = 0,
  SOCmax = 100,
  SOCini = NULL,
  window_days = 1,
  window_start_hour = 0,
  flex_window_hours = 24,
  mc.cores = 2
)
}
\arguments{
\item{opt_data}{tibble, optimization contextual data.
The first column must be named \code{datetime} (mandatory) containing the
date time sequence where the optimization algorithm is applied.
The other columns can be:
\itemize{
\item \code{static}: static power demand (in kW) from other sectors like buildings,
offices, etc.
\item \code{grid_capacity}: maximum imported power from the grid (in kW),
for example the contracted power with the energy company.
\item \code{production}: local power generation (in kW).
This is used when \code{opt_objective = "grid"}.
\item \code{price_imported}: price for imported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_exported}: price for exported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
}}

\item{opt_objective}{character, optimization objective being \code{"grid"} (default) or \code{"cost"}}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{window_days}{integer, number of days to consider as optimization window.}

\item{window_start_hour}{integer, starting hour of the optimization window.}

\item{flex_window_hours}{integer, flexibility window length, in hours.
This optional feature lets you apply flexibility only during few hours from the \code{window_start_hour}.
It must be lower than \code{window_days*24} hours.}

\item{mc.cores}{integer, number of cores to use.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile
}
