% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{add_battery_simple}
\alias{add_battery_simple}
\title{Simple battery profile}
\usage{
add_battery_simple(G, L, Bcap, Bc, Bd, SOCini = 0)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{Bcap}{numeric, capacity of the battery}

\item{Bc}{numeric, maximum charging power}

\item{Bd}{numeric, maximum discharging power}

\item{SOCini}{numeric, required State-of-Charge at the window beginning}
}
\value{
numeric vector
}
\description{
Charging when there is surplus, discharging when there is deficit
}
