% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{denormalize_sessions}
\alias{denormalize_sessions}
\title{Denormalize sessions (from timeslot to datetime)}
\usage{
denormalize_sessions(sessions_norm, start, time_interval)
}
\arguments{
\item{sessions_norm}{tibble, normalized sessions data set}

\item{start}{datetime, start datetime value}

\item{time_interval}{integer, interval of time between time slots (minutes)}
}
\value{
tibble
}
\description{
Denormalize sessions (from timeslot to datetime)
}
