% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_conversion_losses}
\alias{get_conversion_losses}
\title{Losses due to charging/discharging process}
\usage{
get_conversion_losses(power, loss_charge, loss_discharge)
}
\arguments{
\item{power}{numeric vector, being positive when charging and negative when discharging}

\item{loss_charge}{numeric, the charging loss in percentage (\%)}

\item{loss_discharge}{numeric, the discharging loss in percentage (\%)}
}
\value{
numeric vector
}
\description{
Losses due to charging/discharging process
}
