% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{get_energy_cost}
\alias{get_energy_cost}
\title{Get energy cost}
\usage{
get_energy_cost(df)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production}, \code{price_imported} and \code{price_exported}}
}
\value{
tibble
}
\description{
The energy cost (in Euros) based on imported and exported energy
and the corresponding prices.
}
