% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_sessions_demand}
\alias{get_sessions_demand}
\title{Obtain timeseries demand from sessions dataset}
\usage{
get_sessions_demand(sessions, dttm_seq = NULL, by = "Profile", normalized = F)
}
\arguments{
\item{sessions}{tibble, sessions data set in standard format marked by \code{{evprof}} package}

\item{dttm_seq}{sequence of datetime values that will be the datetime variable of the returned time-series data frame}

\item{by}{character, being 'Profile' or 'Session'. When \code{by='Profile'} each column corresponds to an EV user profile.}

\item{normalized}{logical, whether the \code{sessions} datetime columns are time-slot values}
}
\value{
tibble
}
\description{
Obtain timeseries demand from sessions dataset
}
