% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-charging.R
\name{get_setpoints}
\alias{get_setpoints}
\title{Set setpoints for smart charging}
\usage{
get_setpoints(
  sessions_window,
  opt_data,
  profiles_demand,
  opt_objective,
  lambda
)
}
\arguments{
\item{sessions_window}{tibble, sessions corresponding to a single windows}

\item{opt_data}{tibble, optimization data}

\item{profiles_demand}{tibble, user profiles power demand}

\item{opt_objective}{character, optimization objective}

\item{lambda}{numeric, penalty on change for the flexible load.}
}
\description{
Set setpoints for smart charging
}
\keyword{internal}
