% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_storage_level}
\alias{get_storage_level}
\title{Accumulated storage level (energy)}
\usage{
get_storage_level(
  power,
  init = 0,
  charge_eff = 1,
  discharge_eff = 1,
  time_resolution = 60
)
}
\arguments{
\item{power}{numeric vector, being positive when charging and negative when discharging}

\item{init}{numeric, initial storage level (in energy units, not \%)}

\item{charge_eff}{numeric, charging efficiency (from 0 to 1, default to 1)}

\item{discharge_eff}{numeric, discharging efficiency (from 0 to 1, default to 1)}

\item{time_resolution}{numeric, time resolution of the time-series (in minutes)}
}
\value{
numeric vector of energy stored
}
\description{
Accumulated storage level (energy)
}
