% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_storage_level}
\alias{get_storage_level}
\title{Accumulated storage level (energy)}
\usage{
get_storage_level(
  power,
  init = 0,
  charge_eff = 1,
  discharge_eff = 1,
  time_resolution = 60
)
}
\arguments{
\item{power}{numeric vector, being positive when charging and negative when discharging}

\item{init}{numeric, initial storage level (in kWh, not \%)}

\item{charge_eff}{numeric, charging efficiency (from 0 to 1, default to 1)}

\item{discharge_eff}{numeric, discharging efficiency (from 0 to 1, default to 1)}

\item{time_resolution}{numeric, time resolution of the time-series (in minutes)}
}
\value{
numeric vector of energy stored
}
\description{
Each value represents the energy level at the beginning of the time slot,
starting from the provided initial State-of-Charge. When the \code{power}
vector carries \code{charge} and \code{discharge} attributes (as provided by
\code{add_battery_optimization()}), those are used internally to compute the
storage evolution when efficiencies are below 1.
}
