% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_storage_level}
\alias{get_storage_level}
\title{Accumulated storage level}
\usage{
get_storage_level(power, init = 0, loss = 0)
}
\arguments{
\item{power}{numeric vector, being positive when charging and negative when discharging}

\item{init}{numeric, initial storage level}

\item{loss}{numeric, the hourly storage loss in percentage (\%/hour), passed to \code{get_storage_losses} function}
}
\value{
numeric vector
}
\description{
Accumulated storage level
}
