% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexibility.R
\name{minimize_grid_flow}
\alias{minimize_grid_flow}
\title{Minimization of the grid flow}
\usage{
minimize_grid_flow(
  w,
  G,
  LF,
  LS = NULL,
  direction = "forward",
  time_horizon = NULL,
  up_to_G = TRUE,
  grid_cap = NULL,
  window_length = NULL
)
}
\arguments{
\item{w}{numeric between \code{0} and \code{1}, being the weight of the flexibility potential}

\item{G}{numeric vector, being the renewable generation profile}

\item{LF}{numeric vector, being the flexible load profile}

\item{LS}{numeric vector, being the static load profile}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from}

\item{up_to_G}{logical, whether to limit the flexible EV demand up to renewable Generation}

\item{grid_cap}{numeric, grid capacity limit. Configuring this value make sense mainly when \code{w = 0} since \code{w > 0} will apply peak shaving}

\item{window_length}{integer, window length. If \code{NULL}, the window length will be the length of \code{G}}
}
\value{
numeric vector
}
\description{
Minimization of the grid flow
}
