% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minimize_grid_flow_window_osqp}
\alias{minimize_grid_flow_window_osqp}
\title{Minimization of the grid flow (just a window)}
\usage{
minimize_grid_flow_window_osqp(
  w,
  G,
  LF,
  LS = NULL,
  direction = "forward",
  time_horizon = NULL,
  only_above_G = FALSE,
  up_to_G = FALSE,
  grid_capacity = NULL
)
}
\arguments{
\item{w}{numeric between \code{0} and \code{1}, being the weight of the flexibility potential}

\item{G}{numeric vector, being the renewable generation profile}

\item{LF}{numeric vector, being the flexible load profile}

\item{LS}{numeric vector, being the static load profile}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from}

\item{only_above_G}{logical, optimize only the part of flexible load that surpasses Generation.
If all demand is lower than Generation the optimization is skipped.}

\item{up_to_G}{logical, limit the flexible load up to Generation.}

\item{grid_capacity}{numeric, grid maximum power capacity that will limit the maximum optimized demand}
}
\value{
numeric vector
}
\description{
Minimization of the grid flow (just a window)
}
