% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalize_sessions}
\alias{normalize_sessions}
\title{Normalize sessions (from datetime to timeslots)}
\usage{
normalize_sessions(sessions, start, time_interval)
}
\arguments{
\item{sessions}{sessions data set}

\item{start}{datetime, start datetime value}

\item{time_interval}{integer, interval of time between time slots (minutes)}
}
\value{
tibble
}
\description{
Normalize sessions (from datetime to timeslots)
}
