% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimize_battery_window}
\alias{optimize_battery_window}
\title{Battery optimal charging/discharging profile to minimize net power and cost (just a window)}
\usage{
optimize_battery_window(
  G,
  L,
  PE,
  PI,
  PTD,
  PTU,
  Bcap,
  Bc,
  Bd,
  SOCmin,
  SOCmax,
  SOCini,
  import_capacity,
  export_capacity,
  w,
  lambda,
  charge_eff,
  discharge_eff
)
}
\arguments{
\item{G}{numeric vector, being the renewable generation profile}

\item{L}{numeric vector, being the load profile}

\item{PE}{numeric vector, electricity prices for exported energy}

\item{PI}{numeric vector, electricity prices for imported energy}

\item{PTD}{numeric vector, prices for turn-down energy use}

\item{PTU}{numeric vector, prices for turn-up energy use}

\item{Bcap}{numeric, capacity of the battery (NOT in kWh but in energy units according to time resolution)}

\item{Bc}{numeric, maximum charging power (in kW)}

\item{Bd}{numeric, maximum discharging power (in kW)}

\item{SOCmin}{numeric, minimum State-of-Charge of the battery}

\item{SOCmax}{numeric, maximum State-of-Charge of the battery}

\item{SOCini}{numeric, required State-of-Charge at the beginning/end of optimization window}

\item{import_capacity}{numeric vector, grid maximum import power capacity that will limit the maximum charging power}

\item{export_capacity}{numeric vector, grid maximum export power capacity that will limit the maximum discharging power}

\item{w}{numeric, optimization objective weight (\code{w=1} minimizes net power while \code{w=0} minimizes cost)}

\item{lambda}{numeric, penalty on change for the flexible load}

\item{charge_eff}{numeric, battery charging efficiency (from 0 to 1)}

\item{discharge_eff}{numeric, battery discharging efficiency (from 0 to 1)}
}
\value{
numeric vector
}
\description{
Battery optimal charging/discharging profile to minimize net power and cost (just a window)
}
\keyword{internal}
