% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimize_demand}
\alias{optimize_demand}
\title{Optimize a vector of flexible demand}
\usage{
optimize_demand(
  LF,
  opt_data,
  opt_objective = "grid",
  direction = "forward",
  time_horizon = NULL,
  window_length = NULL,
  window_start_hour = 0,
  flex_window_length = window_length,
  LFmax = Inf,
  mc.cores = 2
)
}
\arguments{
\item{LF}{numeric vector, being the flexible load profile (in kW)}

\item{opt_data}{tibble, optimization contextual data.
The first column must be named \code{datetime} (mandatory) containing the
date time sequence where the optimization algorithm is applied.
The other columns can be:
\itemize{
\item \code{static}: static power demand (in kW) from other sectors like buildings,
offices, etc.
\item \code{grid_capacity}: maximum imported power from the grid (in kW),
for example the contracted power with the energy company.
\item \code{production}: local power generation (in kW).
This is used when \code{opt_objective = "grid"}.
\item \code{price_imported}: price for imported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_exported}: price for exported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
}}

\item{opt_objective}{character, optimization objective being \code{"grid"} (default) or \code{"cost"}}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from.
If \code{NULL}, the \code{time_horizon} will be the number of rows of \code{op_data}.}

\item{window_length}{integer or integer vector, window length in time slots (not hours).
If \code{NULL}, the \code{window_length} will be the number of rows of \code{op_data}.
It can also be an integer vector, defining a specific length for every
window across the length of \code{opt_data}. In that case, the sum of all vector values
must be equal to the length of \code{opt_data}.}

\item{window_start_hour}{integer, hour to start the first optimization window.}

\item{flex_window_length}{integer, flexibility window length in time slots (not hours).
This optional feature lets you apply flexibility only during few hours from the start of the window.
It can also be an integer vector, defining a specific length for every
window across the length of \code{opt_data}.
It must be lower than \code{window_length}, if not it is limited to \code{window_length} value.}

\item{LFmax}{numeric, value of maximum power (in kW) of the flexible load \code{LF}}

\item{mc.cores}{integer, number of cores to use.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
numeric vector
}
\description{
Optimize a vector of flexible demand
}
