% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimize_demand}
\alias{optimize_demand}
\title{Optimize a vector of flexible demand}
\usage{
optimize_demand(
  opt_data,
  opt_objective = "grid",
  direction = "forward",
  time_horizon = NULL,
  window_days = 1,
  window_start_hour = 0,
  flex_window_hours = NULL,
  lambda = 0,
  mc.cores = 1
)
}
\arguments{
\item{opt_data}{tibble, optimization contextual data.
The first column must be named \code{datetime} (mandatory) containing the
date time sequence where the optimization algorithm is applied.

The second column must be named \code{flexible} (mandatory), being the
power demand (in kW) vector that will be optimized.

The other columns can be (optional):
\itemize{
\item \code{static}: static power demand (in kW) from other sectors like buildings,
offices, etc.
\item \code{grid_capacity}: maximum imported power from the grid (in kW),
for example the contracted power with the energy company.
\item \code{load_capacity}: maximum power that the \code{flexible} load
can consume (in kW).
\item \code{production}: local power generation (in kW).
This is used when \code{opt_objective = "grid"}.
\item \code{price_imported}: price for imported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_exported}: price for exported energy (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_turn_down}: price for turn-down energy use (€/kWh).
This is used when \code{opt_objective = "cost"}.
\item \code{price_turn_up}: price for turn-up energy use (€/kWh).
This is used when \code{opt_objective = "cost"}.
}}

\item{opt_objective}{character, optimization objective being \code{"grid"} (default),
\code{"cost"} or a value between 0 (cost) and 1 (grid).}

\item{direction}{character, being \code{forward} or \code{backward}. The direction where energy can be shifted}

\item{time_horizon}{integer, maximum number of positions to shift energy from.
If \code{NULL}, the \code{time_horizon} will be the number of rows of \code{op_data}.}

\item{window_days}{integer, number of days to consider as optimization window.}

\item{window_start_hour}{integer, starting hour of the optimization window.}

\item{flex_window_hours}{integer, flexibility window length, in hours.
This optional feature lets you apply flexibility only during few hours from the \code{window_start_hour}.
It must be lower than \code{window_days*24} hours.}

\item{lambda}{numeric, penalty on change for the flexible load.}

\item{mc.cores}{integer, number of cores to use.
Must be at least one, and parallelization requires at least two cores.}
}
\value{
numeric vector
}
\description{
See the formulation of the optimization problems in the
\href{https://resourcefully-dev.github.io/flextools/}{documentation website}.
}
