% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{plot_net_power}
\alias{plot_net_power}
\title{Plot net power interaction with the grid}
\usage{
plot_net_power(
  df,
  original_df = NULL,
  import_capacity = NULL,
  export_capacity = NULL,
  ...
)
}
\arguments{
\item{df}{tibble, with columns \code{datetime}, \code{consumption}, \code{production}}

\item{original_df}{tibble with same columns than \code{df} corresponding to the
original scenario (e.g. without flexibility)}

\item{import_capacity}{numeric, maximum power to import from the grid (in kW)}

\item{export_capacity}{numeric, maximum power to export to the grid (in kW, positive).}

\item{...}{extra arguments to pass to dygraphs::dyOptions function}
}
\value{
dygraphs plot
}
\description{
HTML interactive plot showing the hourly energy cost. Also,
a comparison between the original scenario is done when
\code{original_df} is not \code{NULL}.
}
\examples{
df <- dplyr::select(
  energy_profiles,
  datetime,
  production = solar,
  consumption = building
)
df <- dplyr::slice_head(df, n = 300)
head(df)

plot_net_power(df)

# Build another random building profile
building_variation <- rnorm(nrow(df), mean = 0, sd = 1)
df2 <- dplyr::mutate(df, consumption = consumption + building_variation)
plot_net_power(df2, original_df = df)

}
