% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_charging.R
\name{schedule_sessions}
\alias{schedule_sessions}
\title{Schedule sessions according to optimal setpoint}
\usage{
schedule_sessions(
  sessions_prof,
  setpoint_prof,
  method,
  power_th = 0,
  include_log = F,
  power_min = 3.7
)
}
\arguments{
\item{sessions_prof}{tibble, sessions data set normalized from \code{normalize_sessions} function}

\item{setpoint_prof}{tibble with columns \code{datetime}, \code{timeslot} and rest of columns being the names of user profiles}

\item{method}{character, being \code{postpone} or \code{curtail}}

\item{power_th}{numeric, power threshold accepted from setpoint, in percentage}

\item{include_log}{logical, whether to output the algorithm messages for every user profile and time-slot}

\item{power_min}{numeric, minimum power to charge vehicles using curtailment method}
}
\value{
list of two elements \code{sessions} and \code{log}
}
\description{
Schedule sessions according to optimal setpoint
}
