% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-charging.R
\name{set_responsive}
\alias{set_responsive}
\title{Set \code{Responsive} column in \code{sessions}}
\usage{
set_responsive(
  sessions_window,
  dttm_seq,
  responsive,
  opt_objective,
  time_resolution
)
}
\arguments{
\item{sessions_window}{tibble, sessions corresponding to a single windows}

\item{dttm_seq}{datetime vector}

\item{responsive}{named list with responsive ratios}

\item{opt_objective}{character, optimization objective being \code{"none"}, \code{"curtail"},
\code{"grid"}, \code{"cost"} or a value between 0 (cost) and 1 (grid).}

\item{time_resolution}{numeric, time resolution in minutes}
}
\description{
Set \code{Responsive} column in \code{sessions}
}
\keyword{internal}
