% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexibility.R
\name{smart_charging}
\alias{smart_charging}
\title{Smart charging algorithm}
\usage{
smart_charging(
  sessions,
  fitting_data,
  window_length,
  window_start_hour,
  opt_weights,
  responsive,
  power_th = 0,
  up_to_G = TRUE,
  grid_cap = NULL,
  sort_by_flex = TRUE,
  include_msg = FALSE
)
}
\arguments{
\item{sessions}{tibble, sessions data set}

\item{fitting_data}{tibble, optimization fitting data, first column being \code{datetime}.
The other columns could be \code{solar} (solar generation) and \code{fixed} (static demand from other sectors like buildings, offices, ...).
Only sessions starting within the time sequence of column \code{datetime} will be shifted.}

\item{window_length}{integer, number of data points of the optimization window (not in hours)}

\item{window_start_hour}{integer, hour to start the optimization window. If \code{window_start = 6} the EV sessions are optimized from 6:00 to 6:00.}

\item{opt_weights}{Named list with the optimization weight \code{w} of the \code{minimize_grid_flow} function. The names of the list must exactly match the user profiles names.}

\item{responsive}{Named list with the ratio of sessions responsive to smart charging program for each profile. The names of the list must exactly match the user profiles names.}

\item{power_th}{power threshold from to consider flexibility required}

\item{up_to_G}{logical, whether to limit the flexible EV demand up to renewable Generation}

\item{grid_cap}{numeric, grid capacity limit. Configuring this value make sense mainly when \code{w = 0} since \code{w > 0} will apply peak shaving}

\item{sort_by_flex}{logical, whether to sort the sessions to shift from higher to lower flexibility}

\item{include_msg}{logical, whether to output the algorithm messages for every user profile and time-slot}
}
\value{
a list with two elements: optimization setpoints and coordinated sessions schedule
}
\description{
Smart charging algorithm
}
