% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_charging.R
\name{smart_charging}
\alias{smart_charging}
\title{Smart charging algorithm}
\usage{
smart_charging(
  sessions,
  fitting_data,
  method,
  window_length,
  window_start_hour,
  opt_weights,
  responsive,
  only_above_G = FALSE,
  up_to_G = FALSE,
  power_th = 0,
  include_log = FALSE,
  charging_power_min = 3.7,
  charging_minutes_min = 30
)
}
\arguments{
\item{sessions}{tibble, sessions data set}

\item{fitting_data}{tibble, optimization fitting data, first column being \code{datetime}.
The other columns could be \code{solar} (solar generation) and \code{fixed} (static demand from other sectors like buildings, offices, ...).
Only sessions starting within the time sequence of column \code{datetime} will be shifted.
If columns of \code{fitting_data} are user profiles names, these are used as setpoints and no optimization is performed.}

\item{method}{character, smart charging method being \code{postpone} or \code{curtail}}

\item{window_length}{integer, number of data points of the optimization window (not in hours)}

\item{window_start_hour}{integer, hour to start the optimization window. If \code{window_start = 6} the EV sessions are optimized from 6:00 to 6:00.}

\item{opt_weights}{Named list with the optimization weight \code{w} of the \code{minimize_grid_flow} function. The names of the list must exactly match the user profiles names.}

\item{responsive}{Named list with the ratio of sessions responsive to smart charging program for each profile. The names of the list must exactly match the user profiles names.}

\item{only_above_G}{logical, optimize only the part of flexible load that surpasses Generation.
If all demand is lower than Generation the optimization is skipped.}

\item{up_to_G}{logical, whether to limit the flexible EV demand up to renewable Generation}

\item{power_th}{numeric, power threshold accepted from setpoint, in percentage}

\item{include_log}{logical, whether to output the algorithm messages for every user profile and time-slot}

\item{charging_power_min}{numeric, minimum power to charge vehicles using curtailment method}

\item{charging_minutes_min}{integer, minimum time (in minutes) that the vehicle must be charging before interruptions}
}
\value{
a list with two elements: optimization setpoints and coordinated sessions schedule
}
\description{
Smart charging algorithm
}
